Add missing bzip2 dependency.
The bzip2 library is always required. We are assuming that we
can use the system library on most platforms including windows
and macos. However, bzip is not available in the flatpak
runtime se we need to use a bundled version here.

A better route is probably to add libzip2 to the flatpak manifest
as a module. Left as an exercise.

The best option would of course be to get rid od this dependency...